Move TRD readers/writers into independent lib
- Remove obsolete TRD MCLabel alias
