+++ Detectors/TRD/workflow/io/src/TRDCalibratedTrackletWriterSpec.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "TRDWorkflow/TRDCalibratedTrackletWriterSpec.h"
+#include "TRDWorkflowIO/TRDCalibratedTrackletWriterSpec.h"
 #include "DataFormatsTRD/CalibratedTracklet.h"
 
 #include "DPLUtils/MakeRootTreeWriterSpec.h"
