+++ Detectors/TRD/workflow/src/TRDTrackingWorkflow.cxx
@@ -15,7 +15,7 @@
 #include "Framework/WorkflowSpec.h"
 #include "GlobalTrackingWorkflowReaders/TrackTPCITSReaderSpec.h"
 #include "TPCWorkflow/TrackReaderSpec.h"
-#include "TRDWorkflow/TRDTrackletReaderSpec.h"
+#include "TRDWorkflowIO/TRDTrackletReaderSpec.h"
 #include "TRDWorkflow/TRDTrackletTransformerSpec.h"
 #include "TRDWorkflow/TRDGlobalTrackingSpec.h"
 #include "TRDWorkflow/TRDTrackWriterSpec.h"
